--- a/adonis-typings/request.ts
+++ b/adonis-typings/request.ts
@@ -593,6 +593,7 @@ declare module '@ioc:Adonis/Core/Request' {
     allowMethodSpoofing: boolean
     getIp?: (request: RequestContract) => string
     trustProxy: (address: string, distance: number) => boolean
+    enableAsyncHttpContext?: boolean
   }
 
   /**