@RomainLanz @thetutlage 

This is very much WIP but I open it early to discuss the API and approach taken.