@targos 2020-06-30
The CI fails on Node.js 12.0.0 because AsyncLocalStorage is only available since v12.17.0.