@RomainLanz 2020-06-30
Here's some benchmarks results to show the performance impact we have by activating the `AsyncLocalStorage`.

The first benchmark is a simple `Hello World`.

**Without the AsyncLocalStorage**
```bash
PS D:\Git\targos\adonis-experiments> autocannon http://localhost:3333
Running 10s test @ http://localhost:3333
10 connections

┌─────────┬──────┬──────┬───────┬──────┬─────────┬─────────┬──────────┐
│ Stat    │ 2.5% │ 50%  │ 97.5% │ 99%  │ Avg     │ Stdev   │ Max      │
├─────────┼──────┼──────┼───────┼──────┼─────────┼─────────┼──────────┤
│ Latency │ 1 ms │ 1 ms │ 2 ms  │ 3 ms │ 1.07 ms │ 0.56 ms │ 26.52 ms │
└─────────┴──────┴──────┴───────┴──────┴─────────┴─────────┴──────────┘
┌───────────┬────────┬────────┬─────────┬─────────┬─────────┬─────────┬────────┐
│ Stat      │ 1%     │ 2.5%   │ 50%     │ 97.5%   │ Avg     │ Stdev   │ Min    │
├───────────┼────────┼────────┼─────────┼─────────┼─────────┼─────────┼────────┤
│ Req/Sec   │ 4005   │ 4005   │ 8027    │ 8247    │ 7575.91 │ 1167.24 │ 4004   │
├───────────┼────────┼────────┼─────────┼─────────┼─────────┼─────────┼────────┤
│ Bytes/Sec │ 665 kB │ 665 kB │ 1.33 MB │ 1.37 MB │ 1.26 MB │ 194 kB  │ 665 kB │
└───────────┴────────┴────────┴─────────┴─────────┴─────────┴─────────┴────────┘

Req/Bytes counts sampled once per second.

83k requests in 11.06s, 13.8 MB read
```

**With the AsyncLocalStorage**
```bash
PS D:\Git\targos\adonis-experiments> autocannon http://localhost:3333
Running 10s test @ http://localhost:3333
10 connections

┌─────────┬──────┬──────┬───────┬──────┬─────────┬─────────┬──────────┐
│ Stat    │ 2.5% │ 50%  │ 97.5% │ 99%  │ Avg     │ Stdev   │ Max      │
├─────────┼──────┼──────┼───────┼──────┼─────────┼─────────┼──────────┤
│ Latency │ 1 ms │ 1 ms │ 3 ms  │ 4 ms │ 1.17 ms │ 0.65 ms │ 28.05 ms │
└─────────┴──────┴──────┴───────┴──────┴─────────┴─────────┴──────────┘
┌───────────┬────────┬────────┬─────────┬─────────┬─────────┬────────┬────────┐
│ Stat      │ 1%     │ 2.5%   │ 50%     │ 97.5%   │ Avg     │ Stdev  │ Min    │
├───────────┼────────┼────────┼─────────┼─────────┼─────────┼────────┼────────┤
│ Req/Sec   │ 3017   │ 3017   │ 6163    │ 6219    │ 5760.46 │ 904.61 │ 3017   │
├───────────┼────────┼────────┼─────────┼─────────┼─────────┼────────┼────────┤
│ Bytes/Sec │ 556 kB │ 556 kB │ 1.13 MB │ 1.14 MB │ 1.06 MB │ 166 kB │ 555 kB │
└───────────┴────────┴────────┴─────────┴─────────┴─────────┴────────┴────────┘

Req/Bytes counts sampled once per second.

63k requests in 11.06s, 11.7 MB read
```

We can see with that we loosing `~20k req/s`.

----

In a real application, we may have many `await` statements that slow down the `AsyncLocalStorage`.
Here's another benchmark with many `await`.

```ts
Route.get('/', async () => {
  await asyncFn();
  await asyncFn();
  await asyncFn();
  await asyncFn();
  await asyncFn();
  await asyncFn();
  await asyncFn();
  await asyncFn();
  await asyncFn();
  await asyncFn();
  return { hello: AsyncHttpContext.getContext()?.request.id() || 'NOTHING' }
})

async function asyncFn() {
  await new Promise((resolve) => setImmediate(resolve));
}
```

**Without**
```bash
┌─────────┬──────┬──────┬───────┬──────┬─────────┬─────────┬──────────┐
│ Stat    │ 2.5% │ 50%  │ 97.5% │ 99%  │ Avg     │ Stdev   │ Max      │
├─────────┼──────┼──────┼───────┼──────┼─────────┼─────────┼──────────┤
│ Latency │ 1 ms │ 1 ms │ 2 ms  │ 3 ms │ 1.06 ms │ 0.37 ms │ 17.28 ms │
└─────────┴──────┴──────┴───────┴──────┴─────────┴─────────┴──────────┘
┌───────────┬────────┬────────┬─────────┬─────────┬─────────┬─────────┬────────┐
│ Stat      │ 1%     │ 2.5%   │ 50%     │ 97.5%   │ Avg     │ Stdev   │ Min    │
├───────────┼────────┼────────┼─────────┼─────────┼─────────┼─────────┼────────┤
│ Req/Sec   │ 3881   │ 3881   │ 7807    │ 7975    │ 7367    │ 1131.91 │ 3880   │
├───────────┼────────┼────────┼─────────┼─────────┼─────────┼─────────┼────────┤
│ Bytes/Sec │ 563 kB │ 563 kB │ 1.13 MB │ 1.16 MB │ 1.07 MB │ 164 kB  │ 563 kB │
└───────────┴────────┴────────┴─────────┴─────────┴─────────┴─────────┴────────┘

Req/Bytes counts sampled once per second.

81k requests in 11.06s, 11.7 MB read
```

**With**
```
┌─────────┬──────┬──────┬───────┬──────┬─────────┬─────────┬──────────┐
│ Stat    │ 2.5% │ 50%  │ 97.5% │ 99%  │ Avg     │ Stdev   │ Max      │
├─────────┼──────┼──────┼───────┼──────┼─────────┼─────────┼──────────┤
│ Latency │ 2 ms │ 2 ms │ 4 ms  │ 5 ms │ 2.17 ms │ 0.75 ms │ 21.08 ms │
└─────────┴──────┴──────┴───────┴──────┴─────────┴─────────┴──────────┘
┌───────────┬────────┬────────┬────────┬────────┬─────────┬────────┬────────┐
│ Stat      │ 1%     │ 2.5%   │ 50%    │ 97.5%  │ Avg     │ Stdev  │ Min    │
├───────────┼────────┼────────┼────────┼────────┼─────────┼────────┼────────┤
│ Req/Sec   │ 2159   │ 2159   │ 4499   │ 4651   │ 4114.11 │ 764.39 │ 2159   │
├───────────┼────────┼────────┼────────┼────────┼─────────┼────────┼────────┤
│ Bytes/Sec │ 313 kB │ 313 kB │ 652 kB │ 674 kB │ 596 kB  │ 111 kB │ 313 kB │
└───────────┴────────┴────────┴────────┴────────┴─────────┴────────┴────────┘

Req/Bytes counts sampled once per second.

41k requests in 10.06s, 5.96 MB read
```

We can see that the we are loosing way more req/sec with multiple `await` in the handler.