@thetutlage 2020-07-01
Hey @targos, thanks for the PR, this is going to be a helpful addition.

@RomainLanz @targos Yes, the performance hits are quite visible and if I am not wrong, then with every async/await call, the performance degrades linearly? 

Also I think, the API is currently new and we will see performance improvements in the future right? Or is it something that will never happen, because of the design of async local storage?