@targos 2020-07-03
>  if I am not wrong, then with every async/await call, the performance degrades linearly?

I have not checked if it's linear, but it seems so, yeah. The reason is that for every `await` call, the spec requires multiple promises to be created by the engine. Node.js has to go through async hooks `init` callaback for each of them to do the context propagation. This affects async hooks in general, not only AsyncLocalStorage.

> Also I think, the API is currently new and we will see performance improvements in the future right? Or is it something that will never happen, because of the design of async local storage?

I have looked a bit at the implementation and it seems that performance improvements are possible in the future.

/cc @vdeturckheim do you think I've implemented it correctly on the Adonis side?