@thetutlage 2020-07-06
@targos 

So what you suggest, on how should we roll it? 

For me the biggest advantage is to use the `logger` and `profiler` from the current HTTP request when performing certain actions like SQL queries. But not sure, if it is worth the performance degrade that comes with it