@puzpuzpuz 2020-07-24
The only improvement I can think of from the implementation perspective is wrappers around `http`'s `EventEmitter`s (`req`/`res`), as they might emit events in a different context. Although, I'm not use with Adonis listens to those events and/or allows users to listen for them.